document-processor/pyproject.toml
[project]
name = "document-processor"
version = "0.1.0"
description = "Document processing service for esec"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.18.0",
    "watchdog>=3.0.0",
    "pillow>=10.2.0",
    "pypdf>=3.17.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "python-magic>=0.4.27",
    "duckdb>=0.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
document_processor = ["py.typed"]